refactor: extract shared PRAGMA helper, deduplicate resolve_active_profile, fix desktop migration gap
- Extract `configure_sqlite_connection()` in crates/db to eliminate PRAGMA
  duplication between `initialize_database()` and `open_raw_sqlite_pool()`
- Move `resolve_active_profile()` and `migrate_flat_layout()` from
  services/api/src/main.rs to services/api/src/lib.rs so both CLI and
  desktop binaries share the same implementation
- Add `migrate_flat_layout()` call to desktop startup (was missing — latent
  bug where desktop would never migrate flat layouts)
- Remove stale narrating comment in auth_steps.rs
