+++ apps/desktop/src/lib.rs
@@ -5,7 +5,10 @@ use tokio_util::sync::CancellationToken;
 use tracing_subscriber::EnvFilter;
 
 use mokumo_api::discovery::MdnsHandle;
-use mokumo_api::{ServerConfig, build_app_with_shutdown, discovery, ensure_data_dirs, try_bind};
+use mokumo_api::{
+    ServerConfig, build_app_with_shutdown, discovery, ensure_data_dirs, migrate_flat_layout,
+    resolve_active_profile, try_bind,
+};
 
 const DEFAULT_PORT: u16 = 6565;
 const DEFAULT_HOST: &str = "0.0.0.0";
@@ -50,24 +53,6 @@ fn initial_webview_url(host: &str, port: u16, setup_token: Option<&str>) -> Stri
     format!("http://{host}:{port}{path}")
 }
 
-/// Read the `active_profile` file from the data directory.
-///
-/// Returns `"demo"` if the file does not exist (first launch defaults to demo).
-fn resolve_active_profile(data_dir: &std::path::Path) -> String {
-    let profile_path = data_dir.join("active_profile");
-    match std::fs::read_to_string(&profile_path) {
-        Ok(contents) => {
-            let trimmed = contents.trim().to_string();
-            if trimmed.is_empty() {
-                "demo".to_string()
-            } else {
-                trimmed
-            }
-        }
-        Err(_) => "demo".to_string(),
-    }
-}
-
 /// Initialize the server: create dirs, backup, run migrations, build app, bind port.
 ///
 /// Extracted so the orchestration sequence can be tested without a window system.
@@ -85,6 +70,9 @@ async fn init_server(
     };
 
     ensure_data_dirs(&config.data_dir)?;
+
+    // Migrate flat layout to dual-directory structure (idempotent)
+    migrate_flat_layout(&config.data_dir)?;
 
     // Resolve which profile to use
     let profile = resolve_active_profile(&config.data_dir);
